Pass the query string to route handlers
The query string shouldn’t be used in route *determination*, but in
route *handling* it’s definitely useful. With this, we can move from
things like this:

    {
	url: /^search(?::(.*))?$/,
	handle ( _, keywords ) {
	    this.goSearch( keywords );
	},
    },

Which will handle URLs like https://www.example.com/search:foo; to this:

    {
	url: /^search$/,
	handle ( _, queryParams ) {
	    this.goSearch( queryParams.q );
	},
    },

Which will handle URLs like https://www.example.com/search?q=foo, which
is decidedly more idiomatic on the web.

As with much of the rest of this batch of O.Router changes, this is a
breaking change. All routes that use capturing groups will need to
ignore the new second argument (the query string, which is never null).