Don't remove id <-> sk mapping on destroy
There may still be foreign key references to it in other records, and we may
still receive references to it in deltas from the server. The only sensible
way to clean these up is a GC that scans all data objects for foreign key refs.
Another time.

We do need to make sure we don't return old ids if we undestroy a record though,
so we check the status of the store key before returning the id.