Fix stopping top-transform animation before end
We optimise animating the "top" value by rewriting this to be the "transform"
value, however if you changed the animation mid-flight it would be broken
because this internal optimisation leaked. We now update the "current" styles
as though we were animating as requested, so we can correctly calculate
changes mid-way.