Fix infinite request loop reverting property
If an attribute is unknown on the server, it will often not be returned.
If you try to change that, the update will therefore probably be
rejected by the server. But handling the rejection the client tries to
revert the change, and was failing to do so because it was just writing
back the original object which didn't contain the attribute. It now
checks for this and writes in the default value for the attribute if
necessary.