Update Todo example for Overture, JMAP changes
The RPC language which was a precursor of what is now being standardised
in JMAP has now been replaced with JMAP. This means changing getFoo to
Foo/get and things like that. Multi-account support has necessitated
certain changes in that as well.

Connection.js has been updated to the current state of that file from
the jmap-js project.

fixtures.js has some additional changes in line with what we’ve done in
the equivalent stuff in FastMail and/or Topicbox.

This also includes some style changes required due to DOM structure
changes in ToolbarView and I think ClearSearchButtonView too.

I have also made it so that when served over HTTP it will assume the
presence of an actual JMAP server on the same origin, rather than
loading fixtures.js. This won’t be suitable for the current hosting
arrangement on overturejs.com, but that hasn’t been updated in yonks. We
have such a simple backend, though it wants GUID IDs instead of the
likes of “inbox”. We’ll think more about details of that later.