Check for changes in StoreUndoManager#dataDidChange
The store delays calculating changes so as not to repeat the same work
many times. But the StoreUndoManager#dataDidChange is triggered
immediately by the events fired on the record, so we need to force the
recalculation early to ensure the hasChanges value is correct.

We could alternatively just delay the dataDidChange fn to the after
queue and override saveUndoCheckpoint to call checkForChanges first
(and then remove the same from getUndoData), but that involves
overriding one more method, and I don't think there's much difference in
efficiency, so I went with this more direct approach.