Make `make docs` not abort
It still doesn’t produce fully successful documentation (the TOC on each
page seems to be mostly non-functional), but at least it no longer
aborts.

Shifting Parameters to the end isn’t strictly necessary, but is
consistent with the rest. The real fix was removing the fifth space on
the method. Fun times.