docbuilder: update for ES modules
The header blocks are no more in an ES modules; that information was
used for the module and the filename.

Filename we can obtain without difficulty: just start actually passing
it through, rather than depending only on the contents of the file.

Module is a mite trickier (the concept isn’t especially suitable), but
we can approximate what it meant by just taking the first directory
component of the filename inside the source directory. This yields an
all-lowercase result unlike what we had formerly, but I don’t mind that.