Todo: fix parseSearch.js
Turns out that `{init(){}}` and `{init:function(){}}` are *not*
semantically equivalent after all: in the case of the former, the init
function is not a constructor; that is, it does not have a "prototype"
property. Combined with not specifying a superclass via Extends, O.Class
requires that init be a constructor.

There are two reasonable solutions to such situations:

1. Make SearchTreeNode an O.Object subclass: not necessary, because we
   don’t have or want any observable properties. So skip this.

2. Switch SearchTreeNode to use native ES6 class syntax: definitely the
   desirable solution, and what I did for some similar cases in Overture
   itself.