Todo: refactor views
No more poorlyNamedModule.js. I also decided to shift the root view into
the entry module, as part of fixing its selectNone method to work on
todoListsView as well. In the current structure, it is logically
depending on todosView and todoListsView, so it makes sense to define
the RootView and call the insertView at the same time. Further
refactoring might end up with this decision reverted, too; who can tell?