O.DOMEvent.lookupKey: fix Shift-* on old browsers
On Shift-* events, KeyboardEvent#key is capitalised; unless we wish to
change how we handle *that* (not entirely unreasonable, I may add; I’m
not fond of saying “Shift-Q”, seeing as “Shift-q”, referring to the
physical keystrokes, or “Q”, referring to the logical input, are more
precise), we need to change the fallback for older browsers that don’t
support KeyboardEvent#key, so that they produce an uppercase letter.

I’m not confident about the altAndShift part, because I’m not used to
dealing with such keyboard layouts and so lack a suitable intuition on
how the events come through. I *think* that we don’t want to change it;
at the very least, not changing it will only inconvenience old browsers.