Stop using object shorthand for init method
The reasons are explained in the comment added in core/Core.js.

The source as *written* was broken; the only reason it worked was
because we have Bublé in place in the build process, and it translates
`init () {}` to `init: function () {}` itself, incidentally fixing that
problem. I’m now changing this to fix that discrepancy/bug, in part so
that I can start importing directly from Overture on a project where we
don’t need Bublé.