Make Element.js export items individually
This also requires when and unless to be exported individually from
SwitchView.js.

I would like to kill the O.Element namespace object from the modules,
making it purely an aspect of the source/Overture.js façade, but I can’t
do that until I kill off View#draw’s second and third arguments (Element
and Element.create), leaving it only taking the layer (and remember that
even *that* is not strictly essential).

In the mean time, everything *else* is now using these module imports
directly rather than using the O.Element namespace object.

source/dom/Element.js has a massive diff, but it’s almost entirely just
reducing the level of indentation and then some reordering of the
methods so that `const` can be used instead of that unpleasant `let`
hack that I had near the start of the file.

This is a meaninglessly small performance and code size win, after
processing through Rollup. (Without Rollup, I can’t say whether it’s a
win or not.)