Store: remove the concept of one default account
isDefault is a relic of JMAP’s past, when there was a global default,
rather than a primary account *per specification URI* as it is now.

_defaultAccountId is obsolete and removed.

_skToAccountId will now commonly be larger, having a record for every
record rather than every account except for the default. On the other
hand, life is simpler this way.

There was a fairly serious consistency problem before this commit:
basically, the idea of passing a null accountId in and letting it look
up the default only worked for *the* default account (with jmap-js, that
meant mail data). If you have multiple accounts for different data types
(which we do with FastMail), this was sorely misleading and dangerous. I
believe it could lead to you trying to e.g. add calendar objects to a
mail-only account, rather than using the calendar account.

getDefaultAccountId is renamed to getPrimaryAccountIdForType, now taking
a type argument, which jmap-js in its alternative implementation will be
able to use to look up the proper value through Type.dataGroup and
JMAP.auth.primaryAccounts. The stock implementation leaves O.Store not
implementing the concept of primary accountIds. (Throwing a TypeError
in this way will make it much easier to see where things went wrong in
cases where an accountId is missing; historically, it’d just fail when
you tried to access a property on the result of the getAccount() call.
Now, that’ll only happen if a *bad* accountId was specified.)

getAccount also now takes an optional argument Type, which will look up
the primary account for that type if the accountId is missing. Some
places use that, but others do the getPrimaryAccountIdForType call
themselves, because they need the accountId and not just the account
(which doesn’t contain the accountId).

I suppose I scuppered the dummy empty string accountId along the way
with this (e.g. by changing the null check in getAccount to a truthiness
check like everywhere else, and by ditching the special-casing of the
empty string in addAccount, with that old isDefault flag); I believe
that was a relic of bygone times too, and not a thing to be mourned.