Set null rather than empty list when no observers
We can't delete the property because the prototype may have observers defined.
Setting to null makes it much easier to check whether there are any observers
for a particular key, and in fact some code was already trying to do this.

This commit refactors the code to always add/remove observers via methods on
the Metadata prototype to remove code duplication and avoid error.