Drop O.UA.cssProps: user-select is all we need now
All of the others are supported, unprefixed, in all of the browsers we
care about; we no longer care about IE10, Safari 8, Firefox 21 or Chrome
35 which I think are the oldest versions requiring prefixes for any of
the items (`transform` for Chrome 29–35, `display: flex` for the rest).

user-select is the only one that still needs a prefix. Due to subtleties
in behaviour, it’s still prefixed in IE, Edge (which supports the
-webkit- prefix, which our code will use, as well as -ms-), Firefox and
Safari. Chrome 54 and browsers of that lineage are the only ones
supporting it unprefixed at present.