Shuffle Record.{attr, toMany, toOne} around
Record.toMany was in ToManyAttribute.js, and I forgot about it and
missed shifting its assignment into the façade.

Record.toOne was always in Record.js rather than ToOneAttribute.js, and
Record.attr likewise in Record.js rather than RecordAttribute.js.

I believe these were done where they were because of dependency ordering
in the old build system—Record.js depended on RecordAttribute.js and
ToOneAttribute.js, but not ToManyAttribute.js. But now that they’re
only being added to Record in the façade, that reason disappears, and
they can return to their wonted places.

In all of FastMail and Topicbox, we’ve never actually needed to refer to
ToOneAttribute, ToManyAttribute or RecordAttribute; they only really
need to be referred to in Overture itself—in fact, I wouldn’t mind
making them implementation details, with attr, toMany and toOne being
the only API. Because of that, I decided to rename the modules to those
names rather than the *Attribute.js names, with the corresponding
function as the default export instead of the class.