Improve restoreEncodedState global param handling
Before this it was in practice a bit broken, with restoreEncodedState
expecting encodedState to have the global query params, but most of the
rest expecting *not* to have the global query params.

After this, the global parameters will only be applied from the URL when
you first load the page, or if you deliberately ask for it with this new
applyGlobalParams argument to restoreEncodedState.