Rejig restoreEncodedState for clarity, efficiency
This version is a little shorter and (I believe) clearer. It’s also
going to be a little more efficient when !applyGlobalParams (the typical
case), and possibly a little more efficient when applyGlobalParams.

This also fixes a rather important logic bug that caused non-global
query string parameters to be lost if not applying global parameters.
(The check `globalNames.has( name )` needed to be negated.)