New method Router#stripGlobalParams
Remember when we dropped Router#currentPath as no longer necessary?
Turns out that, in the short term at least, we need something remarkably
similar to it in FastMail. It’s made more interesting by the global
query string parameters. Here’s a function that takes an encoded state
that may include global query params (such as location.href minus
this.baseUrl), and removes the global query params.

I think/hope this method will be able to vanish again sooner or later.
But for now, I think here’s the right place to put it (rather than in
FastMail).