Move when and unless out of O.Element and into O
I’m fed up with Element not being a pure namespace due to when and
unless being (fairly necessarily) defined in SwitchView.js.

So:

- O.Element.when → O.when
- O.Element.unless → O.unless

I could have put them on O.SwitchView, but I see no value in that, and
I’d like to avoid magic in the façade so that it can still be a single
entry point. Precise exports are best for tree-shaking purposes.