Don't destroy record when old id given
If we have an immutable record, an "update" may have actually been a
destroy and create. We may have updated the old record, but the previous
id => sk mapping stays to allow query changes to work. So we need to
check the reverse mapping gives the original id before updating the
store with the destroy.