Prefer submenus displaying to the right of parent
When you have submenus, we were showing them on whichever side of the
parent had more space. But as long as there is sufficient space, they
should actually always display to the right, which is where the arrow
points.

This commit fixes this, using the width of the parent as the required
space to show the submenu on the right, as it's likely to be similar.