Fire observers in order of registration
Previously, we fired events in reverse order of registration so that if
you removed the observer when it fired it would not break the processing
(if you process from start to front, standard iteration would break if
you removed the item you were currently processing as the indexes of
everything still to be processed would change).

With this commit, we no longer mutate the observer array when removing observers
but instead replace it.

With this, we can now fire observers in the order they were added to the
object. This turns out to be very helpful for our model layer at
FastMail: the model adds handlers for when the data for a Type changes
in the store so that it can invalidate a cache, and this needs to happen
before the LocalQuery instances recalculate, which is triggered by the
same firing. We can guarantee that the model cache handlers are added
before any LocalQuery objects are instantiated, so firing in order
ensures the cache is invalidated first as required.

Because we fire the events in order and cache the length before we go, we can
still mutate the array by pushing new observers on the end of it, and these will
not affect the event firing iteration.