Handle /queryChanges with old+new id for same store key
If you have an immutable type, changing something may result in a new id being
issued. We keep track of the mapping from the old id to the store key still so
that we can update remote queries correctly. However, the query changes may
include both the old and new ids in the "removed" array; we need to deduplicate
this to get the correct result.