Fix US week numbering
While Europe, Oceania and Asia generally use ISO8601 week numbering, the rest of
the world takes a different approach. I'm not sure where I got the previous
algorithm from but it seems it's wrong. So I've updated it and added references
to two separate sources for where the new system is from.