Fix O.UA.browser msie detection
I discovered that IE11 was being detected as “other” rather than “msie”!
This is because, although IE10 and earlier had “MSIE” in the user-agent
string, IE11 dropped it, leaving “Trident/7.0” as the thing to detect.

So, I introduced a check for Trident, and dropped the “MSIE” detection
as obsolete, since we don’t support IE10 or earlier any more.