Fix calculating whether we can redo after undo
Undoing some items may not add a redo inverse to the redo stack (e.g. "undo
send", which may not even be possible if it's too late). However, there could
be other items already in the redo stack if the user already did an undo, so
check that to see if we should enable redo.