tools/localise.js: stop sorting by count
Before this change, in code that uses this tool (such as Fastmail’s web
interface), adding a loc() string anywhere triggers what amounts to a
full rebuild, because the loc(string) → loc(number) technique was
sorting by count (so that the most popular string in the whole code base
would be 0, then 1, 2, &c.); this led to a probable negligible
improvement in transferred code size—in the order of a handful of bytes
only.

After this, only adding a loc() call for a *new* string will cause this
rebuild, which is much more tolerable.