Constants for platform-specific keyboard shortcuts
Delete is the first one, and will probably be the only one for the
indefinite future. More could be added if we seek to support more native
keyboard shortcuts.

I’m adding this here as it really is a library-level feature. It’s not
heavy, and dead code removal will be able to strip out if desired.