New eslint rule overture/class-extends
The raison d’être of Class() over ES6 class syntax is that it is
currently necessary to make things like Overture properties work.
But it’s not the only thing required: Extends must be used too, or else
magic properties won’t be properly hooked up.

The solution is that any time Class() is used, it must be justified by
an Extends argument, such as `Extends: Obj`. Anything that doesn’t have
this is a fairly likely candidate for being an error.