Shorthand for SwitchView with more than two views
when() and unless() are *great* shorthand for when you have one or two
values. But several times I’ve come across a situation where I want
three or more alternatives, and this has required me to either
do things another way, or fall back to using the full SwitchView thing,
which is fraught.

So then, this, which is mostly just publishing an existing private API.
It was actually the *name* of the thing that has held me off from
implementing it earlier!

On code style: we’ve preferred this for when() and unless():

    when( … ).show([
        …
    ]).otherwise([
        …
    ]).end()

But for this, I like the look of this:

    choose( … )
    .case( 0, [
        …
    ])
    .case( 1, [
        …
    ])
    .end()

At this time, it still deals in indexes, so a transform function will
typically be necessary; string cases for what might be called enums is a
possibility for the future.

I have extended the eslint rule overture/when-and-unless-usage to cover
this functionality and renamed it overture/switch-view-usage.