Remove duplicate check from metadata.addObserver
It turns out this is a really hot code path, and so anything we can do to make
it faster is worthwhile. There is only one piece of code relying on the current
behaviour as far as I can tell, so let's only check if the observer already
exists here and skip the check everywhere else.