Only check initial list scroll if record changed
The ListKBFocusView ensures that the focused item is in the visible
scroll window. However, as it was checking every time it entered the
document, if you switched to another view and back it could lose your
scroll position, which is annoying. It should only reset the scroll if
the focused item has also changed.