Only allow Cmd/Ctrl-Enter shortcut in input by default
When focused in an input we want to disable any keyboard shortcut that the user
would expect to be an input command. This was previously fairly permissive, but
resulted in behaviour such as the default Mac delete item shortcut
(Cmd-Backspace) still being active, despite this being a standard text editing
command to delete back to start.

With this change, Cmd-Enter and Cmd + an input key (so not arrows, backspace
etc.) are still allowed by default (the most common), but anything else needs
to be opted into explicitly.