Scroll to top on MenuFilterView focus
Rather than scrolling to the top when the first item is focused in
OptionsListView. This fixes the issue where you have multiple list views
and you focus the first item in a lower list: it shouldn't then scroll
all the way to the top as that's not where the view is!