Fix Firefox bug with View Image from cache
Firefox uses the url of the response, not the request, when you use the
"View Image" or "Copy Image Location" actions from the context menu. This means
it may get a URL that no longer works, because the access token used when it
was fetched into the cache is no longer valid. If we essentially clone the
Response object in this case, it will set the response url to null and so use
the request url instead, fixing the problem.

https://bugzilla.mozilla.org/show_bug.cgi?id=1681179