Don't revive expired cache item
Cache items may have a max-last-access time. However, cleanup only
happens after some item has been read from the cache. If the expired
item is the first one to be fetched from cache after it has expired, it
will still be returned because we avoid doing the DB lookup  until after
the fetch. This is OK, but we should still expire it so next time we
fetch it fresh.