Get list item layout from ListView
This removes the duplicate itemHeight property on the ListItemView and
adds functions to convert index <-> offset (pxTop) to the ListView.
These can be overridden to support layouts where items are not fixed
height, or have headings in between.

ListView#itemLayout is a token that gets changed whenever the layout
needs to change to trigger the appropriate redraws.