Fix "flickering" bug on pinch-to-zoom
The contentDidResize method was being called every time we changed the
scale, as this redraws the layout to adjust the view height. This caused
needless extra work and sometimes weird flickering as it fought over the
scale value.

We only actually need to call this if the parent view resizes.