Fix Apple Pencil click support
This method was meant to prevent the root from scrolling, but
inadvertently breaks support for Apple Pencil as the preventDefault on
touchmove stops the click event from happening with the pencil.

Testing again, I think this method is no longer required; in current iOS
the other mitigations seem to be sufficient.