Fix bug processing query delta updates
Finally tracked this down! So, the bug required the following
conditions:
* We had preemptive updates that removed items.
* We got the actual update from the server, which included a removed
  item that was in our list, but not in the preemptive removal. This
  is actually quite rare (requires a concurrent removal from another
  client), which is why the bug was hit rarely.

The logic for calculating the original indexes of removed items iterated
through the *composed* set of removed items, which included those from
the preemptive updates, and included those in the set of removed items
for the normalised update. When you applied this, it would end up
removing more items than it should, causing the query to be out of sync.

When we next fetched a page of results, it could end up with duplicate
items in the array due to this misalignment, causing the list view to
crash.