New EventSource polyfills
The native browser EventSource class does not give enough control; it
can fail to automatically reconnect, you can't detect if the TCP
connection has broken (by timeouts), and you can't see why a connection
failed (e.g. if you get back a 401). It's also not available in
service workers on Firefox.

With modern fetch/readable streams, we can create a much better class
that can be reused in service workers too. Most of this can be made
generic and implemented with XHR, providing full browser coverage with
just two, mostly identical implementations. The main difference with the
XHR one is it will keep all the data in memory until it's closed, while
the streaming one should be very memory efficient.