Handle 'Enter' on keypress to avoid IME conflict
When users type in Asian languages like Japanese, Enter is used to
select from the auto-complete like IME widget. When this happens, we
should not submit the form. The easiest solution is to just check for
Enter on keypress, because IME inputs emit a keydown and keyup, but no keypress.

See https://www.stum.de/2016/06/24/handling-ime-events-in-javascript/
for more.