Remove duplication and fix bug in NestedStore
If you deleted a contact, then undeleted it via "undo" (so it goes via a nested
store) before the server had committed the delete, we would fail to recreate it.
This was because the NestedStore#commitChanges method didn't add it to the
`_created` object in the parent store.

I carefully rechecked all the cases to see if anything else was wrong, and then
found I had already done this before and implemented it correctly as the Store#undestroyRecord method (but at least now I've double checked that and come to the same conclusion!), so actually we should just call that, which also
removes code duplication and means we don't directly fiddle with the parent's
internal data structures. Win!

https://linear.app/fastmail/issue/CLI-538/investigate-accidental-deletion-of-shared-contacts-after-merge-and