Remove futzing with Promise prototype
Now we have dropped IE11 support we have native Promises in all
browsers. This means we can queue a microtask to flush the run loop
queues if needed, and therefore remove the need to modify the Promise
prototype.

This is also more robust in that we'll still flush queues even if we
execute a task outside of the RunLoop.invoke fn.