Update listLength var when setting list length
We cache the list length in a local variable in the
WindowedQuery._applyUpdate method. However, we were not updating it if
we truncated the list, changing its length. This meant that when
we added items to the list, if they were in a position between the old
length and the new length, they would not be inserted at the correct
position. (As the comment says in the file, you can't use splice to
insert beyond the end of the list in the correct place.)

This was causing a crash that was rare, but still our most common
crasher. Finally tracked this down!