Remove timeout from Database opening
We're hitting this on iOS due to its background throttling. It's not
clear that we actually need this at all – as a reference, the idb npm
module used by workbox doesn't do this, and I think I just added it as a
caution when I was working out how this all works.

fastmail-commit: 28ad3cbb7fd05e545862345ae171dc2b7357dd4f