Fix errors in transaction are swallowed
This was a subtle bug that took me a while to understand. The `fn`
passed to the transaction will normally be async. This means it returns
a promise, so if you don't `await` it, any error thrown inside is
implicitly swallowed (you have to add a `.catch` to see it). If we
await the function, it doesn't matter if it's async or not – either way
we'll propagate the error.

fastmail-commit: ad8fcbd5f51a95b4636eee274abe4976cefdfb43