Fix pre-emptive update invalidation
This fixes a bug when applying multiple changes quickly so that one is
still in flight when another is made. This could previously result in
the DIRTY flag being cleared from the query before the change to the
record was committed, causing it to think a pre-emptive update was
incorrect and remove it, and not fetch the real update.

fastmail-commit: 5e8b82ffe00e8bfdf3b9bd6dc11064c2cbce6eae