Translations: don't overwrite manual translations
When we import translations into JMAPUI, translations that have been manually
adjusted may be overwritten if they were adjusted between the time we exported
our dictionary for translation and the time we are merging the translations back
into JMAPUI. This has led to a couple incidents where we fix a translation,
and then inadvertently revert that change, leading to frustration and despair.

To prevent this, we've added the SHA of the HEAD at export time, and compare
the current state of the translation, just prior to applying the update, with
the state of the translation at the time of export. If the translation has
changed during that time, we reject the incoming translation.

This has changed our command line input, and defined a structure to the naming
of our translation CSV files

fastmail-commit: f5c19a9faa3b050c7a17d26924323edb812fad78