Add reference to Firefox limitation bug number
We ran into this issue a long time (10 years?) ago, but it popped up
on Hacker News yesterday by chance with a link to the actual Bugzilla
entry, so might as well put the exact max supported value in plus a link
to Bugzilla. Top marks to cmorgan for correctly deducing the exact
value.

fastmail-commit: e14afff6b2934d9b5184781bb84e51b3a1daba74