Overture: treat scripts in tools as commonjs modules
The recent Overture change to `"type": "module"` causes scripts in the `tools/`
directory to also be treated as ES modules. Introducing a tiny
`tools/package.json` with `"type": "commonjs"` allows these scripts to continue
to be treated as CommonJS modules.

A better long term fix might be to convert these scripts to ES modules, but
this change fixes the immediate problem.

fastmail-commit: 96d77a8a8ce2263d69a497707755030f581ea12a