MenuView: viewMayHaveResized if content changes
Hovering over a MenuOptionView scrolls it into view. It does this by checking the
parent ScrollView's pxHeight. pxLayout is a computed prop with no dependents, and
when we modify the content of the MenuView by clicking "Show All", pxLayout
doesn't change. This resulted in behavior where the list would scroll the selected
option into view, even though the selected option was already fully in view.
By calling MenuController.viewMayHaveResized when the Menu's content changes,
pxLayout (and thereby pxHeight) is recalculated.

fastmail-commit: d1cd5dd408ded948414f9824746f74cf34536cea