MenuButtonView: replicate drag to select on mobile
The menu will open when the user begins to touch the button. An option
can be selected by releasing while over one of the menu items.

File inputs cannot be opened via this method as browsers don't allow
touch events to perform this action. If a file input is selected a
'confirm' event will be fired on the `FileInputView` which can be used
to open confirmation dialog. Interaction with the dialog will allow the
OS file picker interface to be opened.

fastmail-commit: ad6e893abfa4ccb4816e6f661b54c5f905e5272e