Fix overflow menu hiding on Android keyboard open
On Android, when the virtual keyboard is drawn a resize event is
dispatched to the window. The `pxLayout` property on the `RootView` will
then recompute which will notify anything listening to `pxLayout` (or
any derived properties).

If an overflow menu in a `ToolbarView` exists and is open, it will close
itself when the `left` property is recomputed. Since the `left` property
is listening for changes to `pxWidth` and keyboard draws emit resize
events on Android, focusing a `MenuFilterView` input will cause the
popover to immediately close.

fastmail-commit: 0dc90dc95298a8e2fd266ce0766946b90991045b