Fix date parsing in Japanese and similar
In Japanese, you write a character meaning day/month/year after the
appropriate number. We weren't handling this in any way, so parsing
was failing. This fix grabs the full date string for the locale and
strips out the variables, then just treats everything else as a
delimiter. It's a bit of a hack, but does the trick.

fastmail-commit: 8f022d8a5790ca92b21aa247f3fb7e1778d4842a