Fix client drops event source frequently
The navigator.connection.onchange event fires really frequently in
Chromium browsers — not just when you change networks, but when it
thinks the speed may have changed a little, or the RTT. The online
event is actually much more reliable for determining when you change
networks (and it's the only one other browsers support). However, this
is not available in workers in Chrome yet, so we still need to use the
navigator.connection.onchange event there - but only to detect when
we've lost connection, so we don't abort too frequently.

fastmail-commit: 54f33513b3a190c48fbd809c5a54089375248c62