Break after finding kb shortcut to remove
This is slightly more efficient, but more importantly we only ever
remove one registration, which matches that register() always adds a
registration even if one already exists. Otherwise, if there are
multiple registrations, removing one will remove both, which will result
in an empty list, which causes a crash if you try to show the shortcuts
overlay.

fastmail-commit: 5b091d58ceb0c6952333f7b98d25935d7a4a63f2